modules/Module05/pw_04/requirements.txt
# Requirements file for Cloud Composer environment
# Add this to PyPI packages in Cloud Composer

dbt-bigquery==1.7.0
dbt-core==1.7.0
dbt-utils==1.1.1
dbt-expectations==0.10.1
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
pandas==2.0.3
numpy==1.24.3
